application/requirements.txt
coloredlogs
opencv-python
SimpleWebSocketServer
Pillow
ftrobopy
numpy